+++ src/math/transformation/vpExponentialMap.h
@@ -81,10 +81,10 @@ class VISP_EXPORT vpExponentialMap
 public:
   static vpHomogeneousMatrix direct(const vpColVector &v);
   static vpHomogeneousMatrix direct(const vpColVector &v,
-				    const float &delta_t);
+				    const double &delta_t);
   static vpColVector inverse(const vpHomogeneousMatrix &M);
   static vpColVector inverse(const vpHomogeneousMatrix &M,
-			     const float &delta_t);
+			     const double &delta_t);
 
 };
 #endif
